Split update_task_workers.c
