+++ applications/updater/util/update_task_worker_flasher.c
@@ -11,6 +11,8 @@
 #include <toolbox/tar/tar_archive.h>
 #include <toolbox/crc32_calc.h>
 
+#define TAG "UpdWorkerRAM"
+
 #define CHECK_RESULT(x) \
     if(!(x)) {          \
         break;          \
@@ -22,9 +24,6 @@
 #define FLIPPER_ZERO_DFU_DEVICE_CODE 0xFFFF
 /* Time, in ms, to wait for system restart by C2 before crashing */
 #define C2_MODE_SWITCH_TIMEOUT 10000
-
-#define EXT_PATH "/ext"
-#define TAG "UpdWorker"
 
 static const DfuValidationParams flipper_dfu_params = {
     .device = FLIPPER_ZERO_DFU_DEVICE_CODE,
@@ -159,7 +158,7 @@ static bool update_task_write_stack(UpdateTask* update_task) {
         update_task_set_progress(update_task, UpdateTaskStageRadioInstall, 80);
         CHECK_RESULT(ble_glue_fus_wait_operation() == BleGlueCommandResultOK);
         update_task_set_progress(update_task, UpdateTaskStageRadioInstall, 100);
-        // ...system will restart here.
+        /* ...system will restart here. */
         update_task_wait_for_restart(update_task);
         success = true;
     } while(false);
@@ -175,7 +174,7 @@ static bool update_task_remove_stack(UpdateTask* update_task) {
         update_task_set_progress(update_task, UpdateTaskStageRadioErase, 80);
         CHECK_RESULT(ble_glue_fus_wait_operation() == BleGlueCommandResultOK);
         update_task_set_progress(update_task, UpdateTaskStageRadioErase, 100);
-        // ...system will restart here.
+        /* ...system will restart here. */
         update_task_wait_for_restart(update_task);
         success = true;
     } while(false);
@@ -198,7 +197,7 @@ static bool update_task_manage_radiostack(UpdateTask* update_task) {
         bool stack_missing = (c2_state->VersionMajor == 0) && (c2_state->VersionMinor == 0);
 
         if(c2_state->mode == BleGlueC2ModeStack) {
-            // Stack type is not available when we have FUS running.
+            /* Stack type is not available when we have FUS running. */
             bool total_stack_match = stack_version_match &&
                                      (c2_state->StackType == radio_ver->version.type);
             if(total_stack_match) {
@@ -208,47 +207,45 @@ static bool update_task_manage_radiostack(UpdateTask* update_task) {
                 success = true;
                 break;
             } else {
-                // Version or type mismatch. Let's boot to FUS and start updating.
+                /* Version or type mismatch. Let's boot to FUS and start updating. */
                 FURI_LOG_W(TAG, "Restarting to FUS");
                 furi_hal_rtc_set_flag(FuriHalRtcFlagC2Update);
                 CHECK_RESULT(furi_hal_bt_ensure_c2_mode(BleGlueC2ModeFUS));
-                // ...system will restart here.
+                /* ...system will restart here. */
                 update_task_wait_for_restart(update_task);
             }
         } else if(c2_state->mode == BleGlueC2ModeFUS) {
-            // OK, we're in FUS mode.
+            /* OK, we're in FUS mode. */
             update_task_set_progress(update_task, UpdateTaskStageRadioBusy, 10);
             FURI_LOG_W(TAG, "Waiting for FUS to settle");
             ble_glue_fus_wait_operation();
             if(stack_version_match) {
-                // We can't check StackType with FUS, but partial version matches
+                /* We can't check StackType with FUS, but partial version matches */
                 if(furi_hal_rtc_is_flag_set(FuriHalRtcFlagC2Update)) {
-                    // This flag was set when full version was checked.
-                    // And something in versions of the stack didn't match.
-                    // So, clear the flag and drop the stack.
+                    /* This flag was set when full version was checked.
+                     * And something in versions of the stack didn't match.
+                     * So, clear the flag and drop the stack. */
                     furi_hal_rtc_reset_flag(FuriHalRtcFlagC2Update);
                     FURI_LOG_W(TAG, "Forcing stack removal (match)");
                     CHECK_RESULT(update_task_remove_stack(update_task));
                 } else {
-                    // We might just had the stack installed.
-                    // Let's start it up to check its version
+                    /* We might just had the stack installed.
+                     * Let's start it up to check its version */
                     FURI_LOG_W(TAG, "Starting stack to check full version");
                     update_task_set_progress(update_task, UpdateTaskStageRadioBusy, 40);
                     CHECK_RESULT(furi_hal_bt_ensure_c2_mode(BleGlueC2ModeStack));
-                    // ...system will restart here.
+                    /* ...system will restart here. */
                     update_task_wait_for_restart(update_task);
                 }
             } else {
                 if(stack_missing) {
-                    // Install stack.
+                    /* Install stack. */
                     CHECK_RESULT(update_task_write_stack(update_task));
                 } else {
                     CHECK_RESULT(update_task_remove_stack(update_task));
                 }
             }
         }
-
-        //success = true;
     } while(false);
 
     return success;
@@ -261,7 +258,7 @@ bool update_task_validate_optionbytes(UpdateTask* update_task) {
     bool ob_dirty = false;
     const UpdateManifest* manifest = update_task->manifest;
     const FuriHalFlashRawOptionByteData* device_data = furi_hal_flash_ob_get_raw_ptr();
-    for(int32_t idx = 0; idx < FURI_HAL_FLASH_OB_TOTAL_VALUES; ++idx) {
+    for(int32_t idx = 0; idx < (int32_t)FURI_HAL_FLASH_OB_TOTAL_VALUES; ++idx) {
         const uint32_t ref_value = manifest->ob_reference.obs[idx].values.base;
         const uint32_t device_ob_value = device_data->obs[idx].values.base;
         const uint32_t device_ob_value_masked = device_ob_value &
@@ -291,18 +288,12 @@ bool update_task_validate_optionbytes(UpdateTask* update_task) {
 
                 FURI_LOG_W(TAG, "Fixing up OB byte #%d to %08X", idx, patched_value);
                 ob_dirty = true;
-                bool set_success = furi_hal_flash_ob_set_word(idx, patched_value);
 
-                bool is_fixed = set_success &&
+                bool is_fixed = furi_hal_flash_ob_set_word(idx, patched_value) &&
                                 ((device_data->obs[idx].values.base &
                                   manifest->ob_compare_mask.obs[idx].values.base) == ref_value);
 
-                if(is_fixed) {
-                    // restart loop
-                    match = true;
-                    idx = -1;
-                    continue;
-                } else {
+                if(!is_fixed) {
                     /* Things are so bad that fixing what we are allowed to still doesn't match
                      * reference value 
                      */
@@ -365,129 +356,6 @@ int32_t update_task_worker_flash_writer(void* context) {
         furi_hal_rtc_set_boot_mode(FuriHalRtcBootModePostUpdate);
         success = true;
     } while(false);
-
-    return success ? UPDATE_TASK_NOERR : UPDATE_TASK_FAILED;
-}
-
-static bool update_task_pre_update(UpdateTask* update_task) {
-    bool success = false;
-    string_t backup_file_path;
-    string_init(backup_file_path);
-    path_concat(
-        string_get_cstr(update_task->update_path), LFS_BACKUP_DEFAULT_FILENAME, backup_file_path);
-
-    update_task->state.total_stages = 1;
-    update_task_set_progress(update_task, UpdateTaskStageLfsBackup, 0);
-    furi_hal_rtc_set_boot_mode(FuriHalRtcBootModeNormal); // to avoid bootloops
-    if((success = lfs_backup_create(update_task->storage, string_get_cstr(backup_file_path)))) {
-        furi_hal_rtc_set_boot_mode(FuriHalRtcBootModeUpdate);
-    }
-
-    string_clear(backup_file_path);
-    return success;
-}
-
-typedef struct {
-    UpdateTask* update_task;
-    int32_t total_files, processed_files;
-} TarUnpackProgress;
-
-static bool update_task_resource_unpack_cb(const char* name, bool is_directory, void* context) {
-    UNUSED(name);
-    UNUSED(is_directory);
-    TarUnpackProgress* unpack_progress = context;
-    unpack_progress->processed_files++;
-    update_task_set_progress(
-        unpack_progress->update_task,
-        UpdateTaskStageProgress,
-        unpack_progress->processed_files * 100 / (unpack_progress->total_files + 1));
-    return true;
-}
-
-static bool update_task_post_update(UpdateTask* update_task) {
-    bool success = false;
-
-    string_t file_path;
-    string_init(file_path);
-
-    // status text is too long, too few stages to bother with a counter
-    update_task->state.total_stages = 0;
-
-    do {
-        CHECK_RESULT(update_task_parse_manifest(update_task));
-        path_concat(
-            string_get_cstr(update_task->update_path), LFS_BACKUP_DEFAULT_FILENAME, file_path);
-
-        bool unpack_resources = !string_empty_p(update_task->manifest->resource_bundle);
-        if(unpack_resources) {
-            update_task->state.total_stages++;
-        }
-
-        update_task_set_progress(update_task, UpdateTaskStageLfsRestore, 0);
-        furi_hal_rtc_set_boot_mode(FuriHalRtcBootModeNormal);
-
-        CHECK_RESULT(lfs_backup_unpack(update_task->storage, string_get_cstr(file_path)));
-
-        if(unpack_resources) {
-            TarUnpackProgress progress = {
-                .update_task = update_task,
-                .total_files = 0,
-                .processed_files = 0,
-            };
-            update_task_set_progress(update_task, UpdateTaskStageResourcesUpdate, 0);
-
-            path_concat(
-                string_get_cstr(update_task->update_path),
-                string_get_cstr(update_task->manifest->resource_bundle),
-                file_path);
-
-            TarArchive* archive = tar_archive_alloc(update_task->storage);
-            tar_archive_set_file_callback(archive, update_task_resource_unpack_cb, &progress);
-            success = tar_archive_open(archive, string_get_cstr(file_path), TAR_OPEN_MODE_READ);
-            if(success) {
-                progress.total_files = tar_archive_get_entries_count(archive);
-                if(progress.total_files > 0) {
-                    tar_archive_unpack_to(archive, EXT_PATH);
-                }
-            }
-            tar_archive_free(archive);
-        }
-        success = true;
-    } while(false);
-
-    string_clear(file_path);
-    return success;
-}
-
-int32_t update_task_worker_backup_restore(void* context) {
-    furi_assert(context);
-    UpdateTask* update_task = context;
-    bool success = false;
-
-    FuriHalRtcBootMode boot_mode = furi_hal_rtc_get_boot_mode();
-    if((boot_mode != FuriHalRtcBootModePreUpdate) && (boot_mode != FuriHalRtcBootModePostUpdate)) {
-        // no idea how we got here. Clear to normal boot
-        furi_hal_rtc_set_boot_mode(FuriHalRtcBootModeNormal);
-        return UPDATE_TASK_NOERR;
-    }
-
-    update_task->state.current_stage_idx = 0;
-
-    if(!update_operation_get_current_package_path(update_task->storage, update_task->update_path)) {
-        return UPDATE_TASK_FAILED;
-    }
-
-    if(boot_mode == FuriHalRtcBootModePreUpdate) {
-        success = update_task_pre_update(update_task);
-    } else if(boot_mode == FuriHalRtcBootModePostUpdate) {
-        success = update_task_post_update(update_task);
-    }
-
-    if(success) {
-        update_task_set_progress(update_task, UpdateTaskStageCompleted, 100);
-    } else {
-        update_task_set_progress(update_task, UpdateTaskStageError, update_task->state.progress);
-    }
 
     return success ? UPDATE_TASK_NOERR : UPDATE_TASK_FAILED;
 }
