+++ firmware/targets/f7/furi_hal/furi_hal_flash.h
@@ -4,6 +4,25 @@
 #include <stdint.h>
 #include <stddef.h>
 
+#define FURI_HAL_FLASH_OB_RAW_SIZE_BYTES 0x80
+#define FURI_HAL_FLASH_OB_SIZE_WORDS (FURI_HAL_FLASH_OB_RAW_SIZE_BYTES / sizeof(uint32_t))
+#define FURI_HAL_FLASH_OB_TOTAL_VALUES (FURI_HAL_FLASH_OB_SIZE_WORDS / 2)
+
+typedef union {
+    uint8_t bytes[FURI_HAL_FLASH_OB_RAW_SIZE_BYTES];
+    union {
+        struct {
+            uint32_t base;
+            uint32_t complementary_value;
+        } values;
+        uint64_t dword;
+    } obs[FURI_HAL_FLASH_OB_TOTAL_VALUES];
+} FuriHalFlashRawOptionByteData;
+
+_Static_assert(
+    sizeof(FuriHalFlashRawOptionByteData) == FURI_HAL_FLASH_OB_RAW_SIZE_BYTES,
+    "UpdateManifestOptionByteData size error");
+
 /** Init flash, applying necessary workarounds
  */
 void furi_hal_flash_init();
@@ -101,35 +120,25 @@ bool furi_hal_flash_program_page(const uint8_t page, const uint8_t* data, uint16
  */
 int16_t furi_hal_flash_get_page_number(size_t address);
 
-/** Writes whole OB word
+/** Writes OB word, using non-compl. index of register in Flash, OPTION_BYTE_BASE
  *
- * @warning Requires OB to be unlocked
+ * @warning locking operation with critical section, stalls execution
  *
  * @param      word_idx  OB word number
  * @param      value    data to write
  * @return     true if value was written, false for read-only word
  */
-
 bool furi_hal_flash_ob_set_word(size_t word_idx, const uint32_t value);
 
+/** Forces a reload of OB data from flash to registers
+ *
+ * @warning Initializes system restart
+ *
+ */
 void furi_hal_flash_ob_apply();
 
-#define FURI_HAL_FLASH_OB_RAW_SIZE_BYTES 0x80
-#define FURI_HAL_FLASH_OB_SIZE_WORDS (FURI_HAL_FLASH_OB_RAW_SIZE_BYTES / sizeof(uint32_t))
-#define FURI_HAL_FLASH_OB_TOTAL_VALUES (FURI_HAL_FLASH_OB_SIZE_WORDS / 2)
-
-typedef union {
-    uint8_t bytes[FURI_HAL_FLASH_OB_RAW_SIZE_BYTES];
-    union {
-        struct {
-            uint32_t base;
-            uint32_t complementary_value;
-        } values;
-        uint64_t dword;
-    } obs[FURI_HAL_FLASH_OB_TOTAL_VALUES];
-} FuriHalFlashRawOptionByteData;
-_Static_assert(
-    sizeof(FuriHalFlashRawOptionByteData) == FURI_HAL_FLASH_OB_RAW_SIZE_BYTES,
-    "UpdateManifestOptionByteData size error");
-
+/** Get raw OB storage data
+ *
+ * @return     pointer to read-only data of OB (raw + complementary values)
+ */
 const FuriHalFlashRawOptionByteData* furi_hal_flash_ob_get_raw_ptr();
